Rework and simplify voltage calibration
Add code to directly sample only voltage and return vrms based on a
supplied calibration factor.
Add a web server handler to handle handle requests to sense voltage
using the given calibration factor.
Modify the calibration functions in the configuration utility to use the
above code.
Rename and move the graph.htm and edit.htm apps to the SD root.