Change over to using a class to define input channels
Using a class to describe input channels, as opposed to many discrete
arrays of variables, makes the code a lot less cluttered and facilitates
methods to operate on the data.  Also can
be more memory efficient if a large number of channels are possible but
few used.
This is the groundwork for introducing output channels which will also
be represented by a class structure.