Overhaul RTC initialization and power fail detection
With version 4 (prototype) hardware prior to 4.3, the RTC battery backup
worked intermittently (more often not).  Required addition of resistor
to slow the voltage drop to the RTC on power fail.  As part of the
hardware fix, code was added to read the RTC power fail indicator, log
it, and reset.