Merge independant Overeasy changes
Two parallel efforts to achieve the same end.  Fundamentally the same,
but this overlay includes a lot of code cleanup and some rearranging.
Most noticeable is not as many .h files because more functions declared
in IotaWatt.h and local functions declared at the top of their
respectice .cpp files.
Also, IotaWatt.h declares all includes so fewer includes in the other
modules.