use units in script evaluation
Evaluate scripts for the units specified.
With VA now available, it's possible to produce amps, pf as well, so
this change does that.
the run script method in iotaScript changes.  Rather than a callback to
get referenced data, caller provides the current logrec and the previous
logrec.  For real-time evaluation, only a current rec is provided with
the current values.