Add export log
Export log keeps track of import and export as defined by a script (+/-) output and saves the totals in a discrete log.  Have not yet figured out how to make that accessible so no user interface yet.  Log is initiated with a config entry "exportlog".
This effort also introduced changes in how log files are defined and introduces a gl;obal naming convention where globals start with a capital letter.