Prompt on unsaved changes, allow nav if no changes
Resolve confusing UX - if you attempt to click a link on a page with inputs (a cancel button), confusingly nothing would happen. This has been resolved so that if there have been no changes, the user is allowed to navigate away, and if the user has made unsaved changes, the user is simply prompted whether they want to leave the page or not.